+++ test/unit/lib/assets.js
@@ -3,6 +3,7 @@
 var Promise = require('../../../src/lib/promise');
 var loadScript = require('../../../src/lib/assets').loadScript;
 var loadStylesheet = require('../../../src/lib/assets').loadStylesheet;
+var throwIfResolves = require('../../helpers/throw-if-resolves');
 
 describe('assets', function () {
   beforeEach(function () {
@@ -15,22 +16,45 @@ describe('assets', function () {
     beforeEach(function () {
       this.options = {
         id: 'script-id',
-        src: 'script-src'
+        src: 'script-src',
+        container: this.fakeContainer
       };
       this.fakeScriptTag = {
         setAttribute: this.sandbox.stub(),
-        addEventListener: this.sandbox.stub().yieldsAsync()
+        addEventListener: this.sandbox.stub().withArgs('load').yieldsAsync()
       };
 
       this.sandbox.stub(document, 'createElement').returns(this.fakeScriptTag);
     });
 
     it('returns a promise that resolves when script has loaded', function () {
-      expect(loadScript(this.fakeContainer, this.options)).to.be.an.instanceof(Promise);
+      var promise = loadScript(this.options);
+
+      expect(promise).to.be.an.instanceof(Promise);
+
+      return promise;
+    });
+
+    it('rejects when script has errored', function () {
+      this.fakeScriptTag.addEventListener.resetBehavior();
+      this.fakeScriptTag.addEventListener.onCall(1).yieldsAsync();
+
+      return loadScript(this.options).then(throwIfResolves).catch(function (error) {
+        expect(error.message).to.equal('script-src failed to load.');
+      });
+    });
+
+    it('rejects when script has aborted', function () {
+      this.fakeScriptTag.addEventListener.resetBehavior();
+      this.fakeScriptTag.addEventListener.onCall(2).yieldsAsync();
+
+      return loadScript(this.options).then(throwIfResolves).catch(function (error) {
+        expect(error.message).to.equal('script-src has aborted.');
+      });
     });
 
     it('appends a configured script tag to provided container', function () {
-      return loadScript(this.fakeContainer, this.options).then(function () {
+      return loadScript(this.options).then(function () {
         var scriptTag = this.fakeContainer.appendChild.firstCall.args[0];
 
         expect(scriptTag).to.equal(this.fakeScriptTag);
@@ -38,7 +62,7 @@ describe('assets', function () {
         expect(scriptTag.id).to.equal('script-id');
         expect(scriptTag.src).to.equal('script-src');
 
-        expect(scriptTag.addEventListener).to.be.calledOnce;
+        expect(scriptTag.addEventListener).to.be.calledThrice;
         expect(scriptTag.addEventListener).to.be.calledWith('load', this.sandbox.match.func);
       }.bind(this));
     });
@@ -49,7 +73,7 @@ describe('assets', function () {
         foo: 'bar'
       };
 
-      return loadScript(this.fakeContainer, this.options).then(function () {
+      return loadScript(this.options).then(function () {
         expect(this.fakeScriptTag.setAttribute).to.be.calledTwice;
         expect(this.fakeScriptTag.setAttribute).to.be.calledWith('data-log-level', 'warn');
         expect(this.fakeScriptTag.setAttribute).to.be.calledWith('data-foo', 'bar');
@@ -71,7 +95,7 @@ describe('assets', function () {
       loadStylesheet({
         id: 'stylesheet-id',
         href: 'stylesheet-href',
-        head: this.fakeHead
+        container: this.fakeHead
       });
 
       stylesheet = this.fakeHead.appendChild.firstCall.args[0];
@@ -89,7 +113,7 @@ describe('assets', function () {
       loadStylesheet({
         id: 'stylesheet-id-1',
         href: 'stylesheet-href',
-        head: this.fakeHead
+        container: this.fakeHead
       });
 
       stylesheet = this.fakeHead.insertBefore.firstCall.args[0];
@@ -105,7 +129,7 @@ describe('assets', function () {
       loadStylesheet({
         id: 'stylesheet-id-1',
         href: 'stylesheet-href',
-        head: this.fakeHead
+        container: this.fakeHead
       });
 
       stylesheet = this.fakeHead.appendChild.firstCall.args[0];